fix bugs

+++ 0x15-file_io/1-create_file.c
@@ -10,9 +10,9 @@
 
 int create_file(const char *filename, char *text_content)
 {
-	ssize_t fd;
-	ssize_t w;
-	ssize_t length;
+	int fd;
+	int w;
+	int length;
 
 	if (!filename)
 		return (-1);
@@ -34,7 +34,7 @@ int create_file(const char *filename, char *text_content)
 
 	if (w == -1)
 		return (-1);
-		
+
 	close(fd);
 	return (1);
 }
